requirements.txt
flask
flask-smorest
python-dotenv
easyocr
Pillow